The original syntax-highlighting.py needed a couple of tweaks applied in order to get the script working properly in windows.
The main thing to understand was that Msysgit pays no attention to the directory in the shebang path, it only looks at the word after the last slash.  It then uses that word to lookup the exe based on the path variable.  It also ignores any parameters.
So for example, with a shebang of #!/c/python27/python it would find the first python.exe found while searching through the paths in the PATH environmental variable.  This means that it is possible to have a python.exe located in c:\python27\python but not use it or find it.

#Currently the logic that makes this occur is in compat/mingw.c @ parse_interpreter()